Another 64bit-related fix

+++ Objective-Zip/FileInZipInfo.h
@@ -39,22 +39,22 @@
 @interface FileInZipInfo : NSObject {
 	
 @private
-	NSUInteger _length;
+	uint64_t _length;
 	ZipCompressionLevel _level;
 	BOOL _crypted;
-	NSUInteger _size;
+	uint64_t _size;
 	NSDate *_date;
 	NSUInteger _crc32;
 	NSString *_name;
 }
 
-- (id) initWithName:(NSString *)name length:(NSUInteger)length level:(ZipCompressionLevel)level crypted:(BOOL)crypted size:(NSUInteger)size date:(NSDate *)date crc32:(NSUInteger)crc32;
+- (id) initWithName:(NSString *)name length:(uint64_t)length level:(ZipCompressionLevel)level crypted:(BOOL)crypted size:(uint64_t)size date:(NSDate *)date crc32:(NSUInteger)crc32;
 
 @property (nonatomic, readonly) NSString *name;
-@property (nonatomic, readonly) NSUInteger length;
+@property (nonatomic, readonly) uint64_t length;
 @property (nonatomic, readonly) ZipCompressionLevel level;
 @property (nonatomic, readonly) BOOL crypted;
-@property (nonatomic, readonly) NSUInteger size;
+@property (nonatomic, readonly) uint64_t size;
 @property (nonatomic, readonly) NSDate *date;
 @property (nonatomic, readonly) NSUInteger crc32;
